+++ app/interface/blueprints/main/add.py
@@ -13,8 +13,10 @@ from flask import (
 )
 
 from app.utils import *
-from app.interface.forms.views import ValidateID, AddForm
-from app.interface.forms.validation_errors import validation_errors
+#from app.interface.forms.views import ValidateID, AddForm
+#from app.interface.forms.validation_errors import validation_errors
+
+from app.interface.forms import ValidateID, AddForm, validation_errors
 
 from app.container import container