+++ app/interface/blueprints/main/jobs.py
@@ -11,8 +11,10 @@ from flask import (
 )
 
 from app.utils import *
-from app.interface.forms.views import ValidateID, JobsForm
-from app.interface.forms.validation_errors import validation_errors
+#from app.interface.forms.views import ValidateID, JobsForm
+#from app.interface.forms.validation_errors import validation_errors
+
+from app.interface.forms import ValidateID, JobsForm, validation_errors
 
 from . import main_bp