+++ app/interface/blueprints/main/settings.py
@@ -10,8 +10,9 @@ from flask import (
 )
 
 from app.utils import *
-from app.interface.forms.views import SettingsForm
-from app.interface.forms.validation_errors import validation_errors
+#from app.interface.forms.views import SettingsForm
+#from app.interface.forms.validation_errors import validation_errors
+from app.interface.forms import validation_errors, SettingsForm
 
 from app.container import container