+++ app/interface/forms/auth/otp_setup.py
@@ -1,0 +1,14 @@
+from ..base import *
+from ..helpers import ValidateOTPCode
+
+class OTPSetupForm(FlaskForm):
+    user_id = None
+
+    otp_code = IntegerField(
+        validators=[
+            InputRequired(),
+            ValidateOTPCode(),
+        ],
+    )
+
+    submit = SubmitField("Submit")