+++ app/interface/forms/main/download.py
@@ -1,0 +1,6 @@
+from ..base import *
+
+class DownloadForm(FlaskForm):
+    server_id = HiddenField(validators=[InputRequired(), ServerExists()])
+    cfg_path = HiddenField(validators=[InputRequired(), ValidConfigFile()])
+    download_submit = SubmitField("Download Config File")