+++ app/interface/forms/main/save.py
@@ -1,0 +1,7 @@
+from ..base import *
+
+class SaveForm(FlaskForm):
+    server_id = HiddenField(validators=[InputRequired(), ServerExists()])
+    cfg_path = HiddenField(validators=[InputRequired(), ValidConfigFile()])
+    file_contents = TextAreaField(validators=[InputRequired()])
+    save_submit = SubmitField("Save File")