+++ app/interface/forms/main/select_cfg.py
@@ -1,0 +1,5 @@
+from ..base import *
+
+class SelectCfgForm(Form):
+    server_id = HiddenField(validators=[InputRequired(), ServerExists()])
+    cfg_path = HiddenField(validators=[InputRequired(), ValidConfigFile()])